+++ Flappy/flappy.h
@@ -1,0 +1,8 @@
+#pragma once
+
+void reset_flappy(void);
+void initFont(Font *font);
+void init_flappy(void);
+void exit_flappy(void);
+void initBoule(Boule *boule);
+extern Game game;
