+++ Flappy/game.h
@@ -1,5 +1,4 @@
-#ifndef DEF_GAME
-#define DEF_GAME
+#pragma once
 
 void run_flappy(void);
 void updateValues(Boule *boule, Tuyau tuyaux[]);
@@ -8,5 +7,3 @@ void nextTuyau(Tuyau *tuyau, Tuyau *previousTuyau);
 int collision(Boule boule, Tuyau tuyau);
 void initBoule(Boule *boule);
 int move(SDL_Surface *ecran, Boule *boule, Tuyau tuyaux[], int *score);
-
-#endif
